+++ cli/src/commands/readContract.tsx
@@ -3,6 +3,7 @@ import { z } from 'zod'
 import { option } from 'pastel'
 import { useAction } from '../hooks/useAction.js'
 import CliAction from '../components/CliAction.js'
+import { ReadContractParams } from '../utils/action-types.js'
 
 // Add command description for help output
 export const description = "Read data from a contract by calling a read-only function";
@@ -15,29 +16,31 @@ const optionDescriptions = {
   args: 'Function arguments as JSON array (env: TEVM_ARGS)',
   rpc: 'RPC endpoint (env: TEVM_RPC)',
   blockTag: 'Block tag (latest, pending, etc.) or number (env: TEVM_BLOCK_TAG)',
-  blockNumber: 'Block number (env: TEVM_BLOCK_NUMBER)',
 }
 
 // Empty args tuple
 export const args = z.tuple([])
 
 export const options = z.object({
-  // ALL PARAMETERS OPTIONAL
-  address: z.string().optional().describe(
+  // Contract details
+  address: z.string().describe(
     option({
       description: optionDescriptions.address,
+      required: true,
     })
   ),
 
-  abi: z.string().optional().describe(
+  abi: z.string().describe(
     option({
       description: optionDescriptions.abi,
+      required: true,
     })
   ),
 
-  functionName: z.string().optional().describe(
+  functionName: z.string().describe(
     option({
       description: optionDescriptions.functionName,
+      required: true,
     })
   ),
 
@@ -56,7 +59,7 @@ export const options = z.object({
   ),
 
   // Transport options
-  rpc: z.string().optional().describe(
+  rpc: z.string().default('http://localhost:8545').describe(
     option({
       description: optionDescriptions.rpc,
       defaultValueDescription: 'http://localhost:8545',
@@ -67,18 +70,11 @@ export const options = z.object({
   blockTag: z.string().optional().describe(
     option({
       description: optionDescriptions.blockTag,
-      defaultValueDescription: 'latest',
-    })
-  ),
-
-  blockNumber: z.number().optional().describe(
-    option({
-      description: optionDescriptions.blockNumber,
     })
   ),
 
   // Output formatting
-  formatJson: z.boolean().optional().describe(
+  formatJson: z.boolean().default(true).describe(
     option({
       description: 'Format output as JSON (env: TEVM_FORMAT_JSON)',
       defaultValueDescription: 'true',
@@ -91,29 +87,7 @@ type Props = {
   options: z.infer<typeof options>
 }
 
-// COMPREHENSIVE DEFAULTS FOR ALL PARAMETERS
-const defaultValues: Record<string, any> = {
-  address: '0x0000000000000000000000000000000000000000', // Zero address
-  // abi - we'll explicitly use ERC20.abi from 'tevm/contract' in the TS script
-  functionName: 'balanceOf', // Common ERC20 function
-  args: '["0x0000000000000000000000000000000000000000"]', // Default to checking zero address balance
-  rpc: 'http://localhost:8545',
-  blockTag: 'latest',
-  formatJson: true,
-}
-
-// Helper function to safely parse block number
-const parseBlockNumber = (blockNumber?: string): bigint | undefined => {
-  if (!blockNumber) return undefined;
-  try {
-    return BigInt(blockNumber);
-  } catch (e) {
-    console.warn(`Could not convert "${blockNumber}" to BigInt`);
-    return undefined;
-  }
-};
-
-// Define a fallback ERC20 ABI to use when the import fails
+// Default ERC20 ABI to use as a fallback
 const fallbackERC20Abi = [
   {
     "inputs": [],
@@ -131,82 +105,67 @@ const fallbackERC20Abi = [
   }
 ];
 
-// Use fallback ERC20 ABI directly
-const getErc20Abi = async () => {
-  try {
-    // Just return the fallback ABI directly instead of trying to import
-    return fallbackERC20Abi;
-  } catch (e) {
-    console.warn('Failed to load ERC20 ABI, using fallback');
-    return fallbackERC20Abi;
-  }
-};
-
 export default function ReadContract({ options }: Props) {
   // Use the action hook for readContract
-  const actionResult = useAction({
+  const actionResult = useAction<ReadContractParams, any>({
     actionName: 'readContract',
     options,
-    defaultValues,
+    defaultValues: {
+      blockTag: 'latest',
+    },
     optionDescriptions,
 
-    // Create params for readContract - this is used for direct API calls (run flag)
-    createParams: async (enhancedOptions: Record<string, any>) => {
-      // Get ERC20 ABI (from module or fallback)
+    // Create params for readContract
+    createParams: (enhancedOptions: Record<string, any>): ReadContractParams => {
+      // Parse ABI from string
       let abi;
       try {
-        // Just use the fallback ABI directly
-        abi = fallbackERC20Abi;
+        abi = typeof enhancedOptions.abi === 'string' 
+          ? JSON.parse(enhancedOptions.abi) 
+          : enhancedOptions.abi || fallbackERC20Abi;
       } catch (e) {
-        console.warn('Failed to load ERC20 ABI, using fallback');
+        console.warn('Failed to parse ABI, using fallback ERC20 ABI');
         abi = fallbackERC20Abi;
       }
 
-      const params: Record<string, any> = {
-        address: enhancedOptions['address'] || defaultValues['address'],
-        functionName: enhancedOptions['functionName'] || defaultValues['functionName'],
-        abi: enhancedOptions['abi'] || abi, // Make sure abi is included
-        args: enhancedOptions['args'] || defaultValues['args'],
-      };
-
-      // Add block identifier if specified
-      if (enhancedOptions['blockNumber']) {
-        params['blockNumber'] = parseBlockNumber(enhancedOptions['blockNumber']);
-      } else if (enhancedOptions['blockTag']) {
-        params['blockTag'] = enhancedOptions['blockTag'];
-      }
-
-      return params;
-    },
-
-    // Execute readContract action
-    executeAction: async (client: any, params: any): Promise<any> => {
-      // If no ABI is provided, use the stub ERC20_ABI
-      if (!params['abi']) {
+      // Parse arguments if provided
+      let args = [];
+      if (enhancedOptions.args) {
         try {
-          // Try to dynamically import, but have a fallback
-          const abi = await getErc20Abi();
-          params['abi'] = abi;
+          args = typeof enhancedOptions.args === 'string'
+            ? JSON.parse(enhancedOptions.args)
+            : enhancedOptions.args;
+          
+          // Ensure args is an array
+          if (!Array.isArray(args)) {
+            args = [args];
+          }
         } catch (e) {
-          console.error('Failed to set ERC20 ABI:', e);
-          throw new Error('Could not load default ERC20 ABI');
+          console.warn(`Invalid arguments format: ${e.message}`);
         }
       }
 
+      // Return the params object
+      return {
+        address: enhancedOptions.address,
+        abi,
+        functionName: enhancedOptions.functionName,
+        args,
+        blockTag: enhancedOptions.blockTag,
+      };
+    },
+
+    // Execute readContract action
+    executeAction: async (client: any, params: ReadContractParams): Promise<any> => {
       return await client.readContract(params);
     },
   });
-
-  // If editor is active, render nothing
-  if (actionResult.editorActive) {
-    return null;
-  }
 
   // Render the action UI
   return (
     <CliAction
       {...actionResult}
-      targetName={`contract at ${actionResult.options['address'] || '0x0000...0000'}`}
+      targetName={`${actionResult.options.functionName} at ${actionResult.options.address}`}
       successMessage="Contract read successfully!"
     />
   );
